requirements.txt
chardet
pyaml
python-dateutil
pillow
sphinx
django
django-filter
django-cors-headers
django-fullclean
djangorestframework
gunicorn
mysqlclient
beautifulsoup4
requests
